Make candidate node pairs from named list or vector

Convert a named list of one-to-one (named vector) or many-to-one (named list) relations into a two-column table of node pairs. It extracts the vector or list element names and the values, unnesting where necessary.

Usage

as_pairs_from_named(x, names_to = "name", values_to = "value")

Arguments

x

named list or vector. Each item contains a vector or value of source node(s) which map to the target node named.

names_to, values_to

character vector specify the new columns to pass the information in x into.

Value

tibble containing .

Examples

veg_vec <- c(eggplant = "aubergine", zucchini = "courgette")
as_pairs_from_named(veg_vec, "au_eng", "uk_eng")
#> # A tibble: 2 × 2
#>   au_eng   uk_eng   
#>   <chr>    <chr>    
#> 1 eggplant aubergine
#> 2 zucchini courgette

animal_list <- list(MAMM = c("elephant", "whale", "monkey"),
                 REPT = c("lizard", "turtle"),
                 CRUS = c("crab"))
as_pairs_from_named(animal_list, "class", "animal")
#> # A tibble: 6 × 2
#>   class animal  
#>   <chr> <chr>   
#> 1 MAMM  elephant
#> 2 MAMM  whale   
#> 3 MAMM  monkey  
#> 4 REPT  lizard  
#> 5 REPT  turtle  
#> 6 CRUS  crab